Use file URIs for MD links in hover popups for Sublime too
The LSP client for Sublime Text doesn't support the filepath:line format
as it is used for Markdown link destinations in hover results (anymore),
but can handle file URIs with a `#line` suffix, like they are already
used for VS Code. This change allows to open links to other files in
hover popups in Sublime Text too.