Completion for string macros
This implements completions for string macros to allow them to complete
without the initial `at` and trailing `_str`. For example, `artif`
completes to `artifact"` like in the REPL. If the completion mode is set
to `:import` the real name is imported (`using Pkg: at-artifact_str`),
and when the completion mode is `:qualify` the qualified string macro is
inserted (`Pkg.artifact"`).