New server option to replace unused function argument with underscore
This implements a new option, `julia.unusedargument`, which can be set
to either `"type"` (default) or `"underscore"`. This option affects the
replacement offered by the remove-unused-argumentname code action. The
default (`"type"`) is to replace `foo` with `::Any`, and `bar::Int` with
`::Int`. When the option is set to `"underscore"` the offered
replacement is instead `_` and `_::Int`, respectively.