Action for adding SPDX header.
This adds a new action for adding a SPDX short form identifier[0] to the
top of the file. The action is active when cursor is at the first line
of the file.

To find the correct license identifer the action first checks in the
tracked documents for existing identifers. This code path supports any
valid identifer for any license. If no such headers are found, the
action falls back to looking in license files in the current workspace
folders. Identification from the license file is currently done by a
heuristic, but this should be OK since the action is invoked by a human,
and not done automatically. Currently only MIT license can be detected
through license files.

[0]: https://spdx.dev/ids/