Simplify JuliaFormatter config search
The previous method of searching for a `.JuliaFormatter.toml` file started in the document's directory, recursing up the filesystem hierarchy until the workspace directory name was not a prefix of the current directory name.

This is problematic because it means that users can no longer have a "global" configuration for all of their Julia projects.

Assume the user has a directory housing all of their Julia projects at `~/Projects/Julia`, and they are currently working on file `~/Projects/Julia/ProjectA/src/ProjectA.jl`. The original behavior would check the following files in order for the presence of a `.JuliaFormatter.toml` file:

- `~/Projects/Julia/ProjectA/src/.JuliaFormatter.toml`
- `~/Projects/Julia/ProjectA/.JuliaFormatter.toml`

However, the user may have actually wanted the algorithm to detect their "global" configuration located at `~/Projects/Julia/.JuliaFormatter.toml`. The previous algorithm would never reach this far up the directory hierarchy, because the base case for terminating the recursion stopped when the workspace directory string (`~/Projects/Julia/ProjectA`) was not a prefix of the currently traversed directory (`~/Projects/Julia`), thus stopping the recursion and returning `nothing`, which forces the loading and parsing of the default configuration.

With this new method, we simply remove the workspace prefix check, simplifying the code in the process. This allows users to place a `.JuliaFormatter.toml` file anywhere up the directory tree from the project they are working on, just as `JuliaFormatter.jl` itself functions.