Unify formatting error codes
Since this is basically the same errors, unify the two
error codes into one. Technically the two are subtly different
(one is trying to format the whole document, while the other
is only trying to format a section), but the underlying cause
is likely the same, so it's fine to merge them. If necessary/required,
they can be disentangled later on too, as currently noone seems to
check them anyway.